Move WebDriver definitions to `wd` module
